requirements.txt
asgiref==3.7.2
bleach==6.1.0
diff-match-patch==20181111
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.3
django-quill-editor==0.1.40
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
packaging==23.2
Pillow==10.0.1
protobuf==4.21.12
PyJWT==2.8.0
python-decouple==3.6
pytz==2023.3.post1
quill-delta==1.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
waitress==2.1.2
webencodings==0.5.1
whitenoise==6.5.0
django-tinymce==3.6.1
dj-database-url==1.0.0
psycopg2-binary==2.9.5
python-decouple==3.6